Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.2",
    "langgraph>=1.0.3",
    "openai>=1.58.1",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]
